core: drop two more redundant u32 intermediate casts post-BoolToUInt fold
After the BoolToUInt removal, two more sites surfaced where an
explicit `static_cast<u32>` widening intermediate sits in front of
a value-preserving cast to a wider type.

gpu_commands.cpp:307 (HandleRenderPolygonCommand)
  Was: static_cast<TickCount>(static_cast<u32>(s_setup_time[a][b][c]))
  s_setup_time is constexpr u16[2][2][2]; widens to s32 (TickCount)
  value-preservingly. The u32 step is a no-op leftover from the
  ZeroExtend32(u16) fold.

pad.h:99 (GetTransferTicks)
  Was: static_cast<TickCount>(static_cast<u32>(m_JOY_BAUD) * 8)
  m_JOY_BAUD is u16; `u16 * 8` integer-promotes to int (= s32),
  fits trivially (max 0x7FFF8). The u32 step adds nothing.

Codegen verified byte-identical at -O2 (single movzwl on x86-64
in both cases, with and without the cast).

+++ src/core/gpu_commands.cpp
@@ -304,7 +304,7 @@ bool GPU::HandleRenderPolygonCommand()
 
   // setup time
   static constexpr u16 s_setup_time[2][2][2] = {{{46, 226}, {334, 496}}, {{82, 262}, {370, 532}}};
-  const TickCount setup_ticks = static_cast<TickCount>(static_cast<u32>(s_setup_time[static_cast<u8>(rc.quad_polygon)][static_cast<u8>(rc.shading_enable)][static_cast<u8>(rc.texture_enable)]));
+  const TickCount setup_ticks = static_cast<TickCount>(s_setup_time[static_cast<u8>(rc.quad_polygon)][static_cast<u8>(rc.shading_enable)][static_cast<u8>(rc.texture_enable)]);
   AddCommandTicks(setup_ticks);
 
   // set draw state up
